EP-813. Renamed stochastic_parameter_factory to stochastic_parameter_builder.
