+++ tvb_epilepsy/base/model/statistical_models/statistical_model.py
@@ -2,7 +2,7 @@ from tvb_epilepsy.base.constants.model_inversion_constants import SIG_EQ_DEF, X1
 from tvb_epilepsy.base.utils.log_error_utils import raise_value_error
 from tvb_epilepsy.base.utils.data_structures_utils import formal_repr, sort_dict
 #TODO: avoid service imported in model
-from tvb_epilepsy.service.stochastic_parameter_factory import set_parameter
+from tvb_epilepsy.service.stochastic_parameter_builder import set_parameter
 
 
 class StatisticalModel(object):
