+++ tvb_epilepsy/service/model_inversion/sde_model_inversion_service.py
@@ -3,7 +3,7 @@ import numpy as np
 from tvb_epilepsy.base.constants.model_constants import model_noise_intensity_dict
 from tvb_epilepsy.base.constants.model_inversion_constants import X1EQ_MAX, SIG_EQ_DEF, SIG_INIT_DEF, SIG_DEF
 from tvb_epilepsy.base.model.statistical_models.sde_statistical_model import SDEStatisticalModel
-from tvb_epilepsy.service.stochastic_parameter_factory import set_parameter_defaults
+from tvb_epilepsy.service.stochastic_parameter_builder import set_parameter_defaults
 from tvb_epilepsy.service.epileptor_model_factory import AVAILABLE_DYNAMICAL_MODELS_NAMES, EPILEPTOR_MODEL_NVARS
 from tvb_epilepsy.service.model_inversion.ode_model_inversion_service import ODEModelInversionService
