+++ tvb_epilepsy/tests/test_plotter.py
@@ -1,14 +1,12 @@
 import os
 import numpy
-
-from tvb_epilepsy.base.model.model_configuration import ModelConfiguration
 from tvb_epilepsy.plot.plotter import Plotter
 from tvb_epilepsy.io.h5_reader import H5Reader
 from tvb_epilepsy.base.model.vep.head import Head
 from tvb_epilepsy.base.model.vep.sensors import Sensors
 from tvb_epilepsy.base.model.vep.surface import Surface
 from tvb_epilepsy.base.constants.configurations import FOLDER_FIGURES, DATA_TEST, FOLDER_LOGS, FOLDER_RES
-from tvb_epilepsy.service.stochastic_parameter_factory import set_parameter
+from tvb_epilepsy.service.stochastic_parameter_builder import set_parameter
 
 head_dir = "head2"
 
@@ -47,9 +45,6 @@ class TestPlotter():
         assert os.path.exists(os.path.join(FOLDER_FIGURES, filename1))
         assert os.path.exists(os.path.join(FOLDER_FIGURES, filename2))
         assert os.path.exists(os.path.join(FOLDER_FIGURES, filename3))
-
-    # def test_plot_vector(self):
-    #     pass
 
     def test_plot_stochastic_parameter(self):
         K_mean = 10 * 2.5 / 87
